cambio de Long a Integer en sus propiedades

+++ src/main/java/com/progra3/proyecto/service/IVehiculoService.java
@@ -6,8 +6,8 @@ import com.progra3.proyecto.entity.Vehiculo;
 
 public interface IVehiculoService {
     List<Vehiculo> getAll();
-    Vehiculo getById(Long id);
+    Vehiculo getById(Integer id);
     Vehiculo save(Vehiculo vehiculo);
-    void delete(Long id);
-    boolean exists(Long id);
+    void delete(Integer id);
+    boolean exists(Integer id);
   }
